(cfbolz, arigo)
Decided that this code is too brittle.  We found two kinds of
bugs in it: bugs that were already there, and bugs that arose
because of TypeCells.

* for the 1st kind of bug: just because the map of an object
doesn't change does not mean that getdictvalue() returns from
the same spot (case of DevolvedDictTerminator); and in that case,
getdictvalue() can actually have random side-effects like changing
version_tag.

* for the 2nd kind of bug: the point is that w_type.lookup() can
now return a different result without the version_tag changing.