first stab at supporting 64 bit windows. This is quite a hack and not meant to be final.
Problem:
Windows 64 bit uses 32 bit integers. PyPy has the implicit
assumption that an inr can hold an address.

Principle:
sys.maxint is patched to be 64 bit. Then the rule about integers
is lifted to not depend on the size of the CPython integer.
Instead, (int, long) is allowed. All overflow checks are (must be)
changed to check an explicit range.

This is work in progress, just the proof of concept.

--HG--
branch : win64 test