Change a detail: when running a JIT-generated piece of code with shadowstack, we used to push two words on the shadowstack, namely a MARKER followed by the address of the frame.  Now we push only one word, which points inside the frame to a MARKER, chosen such that it's clearly a MARKER and not possibly a valid header of a GC object.
The advantage of doing this will become apparent in the following
checkin: the shadowstack can be split at any place, and we don't
have to worry about keeping a MARKER word together with a frame
address.  That would be hard to test and (in case we did it
wrong) would lead to rare bugs.

--HG--
branch : stacklet