Fix issue #1492.
Manual revert in socket.py.  This re-introduces the change done in 2006
in 9db062bb60e3 (anders, implementing explicit refcounting on the
underlying _socket.socket objects).  It reverts a later change where, I
think for goals of being more Python3-like, the refcounter was moved to
the class socket._socketobject.  However this later change is less
compatible with CPython 2.x, as shown by eventlet.