Unroll a (customisable) fixed number of iterations of recursive functions.
In essence, we count how instances of the function we're about to call are
already on the meta-interpreter stack and only stop tracing when that count has
been exceeded. Initial experiments suggest that 7 is a reasonable number, though
this shouldn't be considered fixed in stone, as it's heavily dependent on what
benchmarks one uses.

--HG--
branch : recursion_and_inlining